Refactor undefined event
